# docs/swagger.yaml
definitions:
  admin.ListProjectsResponse:
    properties:
      data:
        $ref: '#/definitions/admin.ListProjectsResponseData'
      error_msg:
        type: string
    type: object
  admin.ListProjectsResponseData:
    properties:
      results:
        items:
          $ref: '#/definitions/admin.ListProjectsResponseDataResult'
        type: array
      total:
        type: integer
    type: object
  admin.ListProjectsResponseDataResult:
    properties:
      description:
        type: string
      id:
        type: string
      nickname:
        type: string
      status:
        $ref: '#/definitions/project.ProjectStatus'
      tags:
        items:
          type: string
        type: array
      username:
        type: string
    type: object
  admin.ReviewProjectRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/project.ProjectStatus'
        enum:
        - 0
        - 1
        - 2
    type: object
  admin.ReviewProjectResponse:
    properties:
      data: {}
      error_msg:
        type: string
    type: object
  admin.listUsersResponse:
    properties:
      data:
        properties:
          total:
            type: integer
          users:
            items:
              $ref: '#/definitions/oauth.User'
            type: array
        type: object
      error_msg:
        type: string
    type: object
...
  oauth.TrustLevel:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    format: int32
    type: integer
    x-enum-varnames:
    - TrustLevelNewUser
    - TrustLevelBasicUser
    - TrustLevelUser
    - TrustLevelActiveUser
    - TrustLevelLeader
  oauth.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_admin:
        type: boolean
      last_login_at:
        type: string
      nickname:
        type: string
      score:
        type: integer
      trust_level:
        $ref: '#/definitions/oauth.TrustLevel'
      updated_at:
        type: string
      username:
        type: string
      violation_count:
        type: integer
    type: object
...
  project.ProjectStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - ProjectStatusNormal
    - ProjectStatusHidden
    - ProjectStatusViolation
...
info:
  contact: {}
  title: LINUX DO CDK
  version: 0.1.0
paths:
  /api/v1/admin/projects:
    get:
      parameters:
      - in: query
        minimum: 1
        name: current
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        - 2
        format: int32
        in: query
        name: status
        type: integer
        x-enum-varnames:
        - ProjectStatusNormal
        - ProjectStatusHidden
        - ProjectStatusViolation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.ListProjectsResponse'
      tags:
      - admin
  /api/v1/admin/projects/{id}/review:
    put:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: string
      - description: 项目信息
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/admin.ReviewProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.ReviewProjectResponse'
      tags:
      - admin
  /api/v1/admin/users:
    get:
      parameters:
      - in: query
        minimum: 1
        name: current
        type: integer
      - in: query
        name: is_active
        type: boolean
      - in: query
        name: is_admin
        type: boolean
      - in: query
        name: min_violations
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        format: int32
        in: query
        name: trust_level
        type: integer
        x-enum-varnames:
        - TrustLevelNewUser
        - TrustLevelBasicUser
        - TrustLevelUser
        - TrustLevelActiveUser
        - TrustLevelLeader
      - in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.listUsersResponse'
      tags:
      - admin
...
swagger: "2.0"